requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
psycopg2-binary
qdrant-client
pydantic
python-dotenv
insightface
onnxruntime
python-multipart